+++ src/app/app.component.ts
@@ -7,6 +7,8 @@ import { FormControl } from '@angular/forms';
   styleUrls: ['./app.component.css']
 })
 export class AppComponent {
+  public phone: number = 123456789;
+  public customMaska: [string, pattern];
 
   public form: FormControl;
   public form1: FormControl;
@@ -15,6 +17,11 @@ export class AppComponent {
   public clearIfNotMatch: FormControl;
   public numberOrStringForm: FormControl;
   public sufixForm: FormControl;
+
+  public pattern: pattern =  {
+    'P': {
+        pattern: new RegExp('\\d'),
+    }};
 
   public numberOrStringFormModel: string | number = '';
   public clearIfNotMatchModel: string | number = '';
@@ -34,6 +41,14 @@ export class AppComponent {
     this.clearIfNotMatch = new FormControl();
     this.numberOrStringForm = new FormControl();
     this.sufixForm = new FormControl('');
+
+    this.customMaska = ['PPP-PPP-PPP', this.pattern];
   }
 
 }
+
+type pattern = {
+    [character: string]: {
+        pattern: RegExp
+    }
+};